# lushop/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/admin/users:
        get:
            tags:
                - UserAdmin
            description: 管理员查看用户列表
            operationId: UserAdmin_ListUsers
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: role
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/lushop.lushop.v1.ListUsersReply'
    /api/admin/users/{id}:
        delete:
            tags:
                - UserAdmin
            description: 管理员踢出用户
            operationId: UserAdmin_KickUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/lushop.lushop.v1.KickUserReply'
...
    /userauth/v1/sms/send:
        post:
            tags:
                - UserAuth
            description: |-
                ========== 短信相关 ==========
                 发送短信验证码
            operationId: UserAuth_SendSms
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/lushop.userauth.v1.SendSmsReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/lushop.userauth.v1.SendSmsReply'
    /userauth/v1/token/issue:
        post:
            tags:
                - UserAuth
            description: |-
                ========== Token 管理 ==========
                 签发 Token（用于登录成功后）
            operationId: UserAuth_IssueToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/lushop.userauth.v1.IssueTokenReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/lushop.userauth.v1.TokenReply'
    /userauth/v1/token/refresh:
        post:
            tags:
                - UserAuth
            description: 刷新 Token
            operationId: UserAuth_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/lushop.userauth.v1.RefreshTokenReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/lushop.userauth.v1.TokenReply'
    /userauth/v1/token/revoke:
        post:
            tags:
                - UserAuth
            description: 撤销 Token（登出）
            operationId: UserAuth_RevokeToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/lushop.userauth.v1.RevokeTokenReq'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
components:
    schemas:
...
        lushop.lushop.v1.KickUserReply:
            type: object
            properties:
                success:
                    type: boolean
        lushop.lushop.v1.ListUsersReply:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/lushop.lushop.v1.UserDetailResponse'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        lushop.lushop.v1.UserDetailResponse:
            type: object
            properties:
                id:
                    type: string
                mobile:
                    type: string
                nickName:
                    type: string
                birthday:
                    type: string
                gender:
                    type: string
                role:
                    type: integer
                    format: int32
            description: user Detail returned
...
        lushop.userauth.v1.IssueTokenReq:
            type: object
            properties:
                userId:
                    type: string
                nickName:
                    type: string
                role:
                    type: integer
                    format: int32
            description: ========== Token 消息 ==========
        lushop.userauth.v1.RefreshTokenReq:
            type: object
            properties:
                refreshToken:
                    type: string
        lushop.userauth.v1.RevokeTokenReq:
            type: object
            properties:
                userId:
                    type: string
        lushop.userauth.v1.SendSmsReply:
            type: object
            properties:
                success:
                    type: boolean
        lushop.userauth.v1.SendSmsReq:
            type: object
            properties:
                mobile:
                    type: string
            description: ========== 短信消息 ==========
        lushop.userauth.v1.TokenReply:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
                expiredAt:
                    type: string
...