# service/userauth/api/userauth/v1/user_auth_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: userauth/v1/user_auth.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationUserAuthGetCaptcha = "/lushop.userauth.v1.UserAuth/GetCaptcha"
const OperationUserAuthIssueToken = "/lushop.userauth.v1.UserAuth/IssueToken"
const OperationUserAuthRefreshToken = "/lushop.userauth.v1.UserAuth/RefreshToken"
const OperationUserAuthRevokeToken = "/lushop.userauth.v1.UserAuth/RevokeToken"
const OperationUserAuthSendSms = "/lushop.userauth.v1.UserAuth/SendSms"

type UserAuthHTTPServer interface {
	// GetCaptcha ========== 验证码相关 ==========
	// 获取图形验证码
	GetCaptcha(context.Context, *emptypb.Empty) (*CaptchaReply, error)
	// IssueToken ========== Token 管理 ==========
	// 签发 Token（用于登录成功后）
	IssueToken(context.Context, *IssueTokenReq) (*TokenReply, error)
	// RefreshToken 刷新 Token
	RefreshToken(context.Context, *RefreshTokenReq) (*TokenReply, error)
	// RevokeToken 撤销 Token（登出）
	RevokeToken(context.Context, *RevokeTokenReq) (*emptypb.Empty, error)
	// SendSms ========== 短信相关 ==========
	// 发送短信验证码
	SendSms(context.Context, *SendSmsReq) (*SendSmsReply, error)
}

func RegisterUserAuthHTTPServer(s *http.Server, srv UserAuthHTTPServer) {
	r := s.Route("/")
	r.GET("/userauth/v1/captcha", _UserAuth_GetCaptcha0_HTTP_Handler(srv))
	r.POST("/userauth/v1/sms/send", _UserAuth_SendSms0_HTTP_Handler(srv))
	r.POST("/userauth/v1/token/issue", _UserAuth_IssueToken0_HTTP_Handler(srv))
	r.POST("/userauth/v1/token/refresh", _UserAuth_RefreshToken0_HTTP_Handler(srv))
	r.POST("/userauth/v1/token/revoke", _UserAuth_RevokeToken0_HTTP_Handler(srv))
}

func _UserAuth_GetCaptcha0_HTTP_Handler(srv UserAuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAuthGetCaptcha)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetCaptcha(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CaptchaReply)
		return ctx.Result(200, reply)
	}
}

func _UserAuth_SendSms0_HTTP_Handler(srv UserAuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendSmsReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAuthSendSms)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendSms(ctx, req.(*SendSmsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendSmsReply)
		return ctx.Result(200, reply)
	}
}

func _UserAuth_IssueToken0_HTTP_Handler(srv UserAuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in IssueTokenReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAuthIssueToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.IssueToken(ctx, req.(*IssueTokenReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TokenReply)
		return ctx.Result(200, reply)
	}
}

func _UserAuth_RefreshToken0_HTTP_Handler(srv UserAuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RefreshTokenReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAuthRefreshToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RefreshToken(ctx, req.(*RefreshTokenReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TokenReply)
		return ctx.Result(200, reply)
	}
}

func _UserAuth_RevokeToken0_HTTP_Handler(srv UserAuthHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RevokeTokenReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationUserAuthRevokeToken)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RevokeToken(ctx, req.(*RevokeTokenReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type UserAuthHTTPClient interface {
	// GetCaptcha ========== 验证码相关 ==========
	// 获取图形验证码
	GetCaptcha(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *CaptchaReply, err error)
	// IssueToken ========== Token 管理 ==========
	// 签发 Token（用于登录成功后）
	IssueToken(ctx context.Context, req *IssueTokenReq, opts ...http.CallOption) (rsp *TokenReply, err error)
	// RefreshToken 刷新 Token
	RefreshToken(ctx context.Context, req *RefreshTokenReq, opts ...http.CallOption) (rsp *TokenReply, err error)
	// RevokeToken 撤销 Token（登出）
	RevokeToken(ctx context.Context, req *RevokeTokenReq, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// SendSms ========== 短信相关 ==========
	// 发送短信验证码
	SendSms(ctx context.Context, req *SendSmsReq, opts ...http.CallOption) (rsp *SendSmsReply, err error)
}

type UserAuthHTTPClientImpl struct {
	cc *http.Client
}

func NewUserAuthHTTPClient(client *http.Client) UserAuthHTTPClient {
	return &UserAuthHTTPClientImpl{client}
}

// GetCaptcha ========== 验证码相关 ==========
// 获取图形验证码
func (c *UserAuthHTTPClientImpl) GetCaptcha(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*CaptchaReply, error) {
	var out CaptchaReply
	pattern := "/userauth/v1/captcha"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationUserAuthGetCaptcha))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// IssueToken ========== Token 管理 ==========
// 签发 Token（用于登录成功后）
func (c *UserAuthHTTPClientImpl) IssueToken(ctx context.Context, in *IssueTokenReq, opts ...http.CallOption) (*TokenReply, error) {
	var out TokenReply
	pattern := "/userauth/v1/token/issue"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserAuthIssueToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RefreshToken 刷新 Token
func (c *UserAuthHTTPClientImpl) RefreshToken(ctx context.Context, in *RefreshTokenReq, opts ...http.CallOption) (*TokenReply, error) {
	var out TokenReply
	pattern := "/userauth/v1/token/refresh"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserAuthRefreshToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// RevokeToken 撤销 Token（登出）
func (c *UserAuthHTTPClientImpl) RevokeToken(ctx context.Context, in *RevokeTokenReq, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/userauth/v1/token/revoke"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserAuthRevokeToken))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendSms ========== 短信相关 ==========
// 发送短信验证码
func (c *UserAuthHTTPClientImpl) SendSms(ctx context.Context, in *SendSmsReq, opts ...http.CallOption) (*SendSmsReply, error) {
	var out SendSmsReply
	pattern := "/userauth/v1/sms/send"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationUserAuthSendSms))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}